Rename to project_admin_service
