+++ tests/server/unit/services/test_project_admin_service.py
@@ -1,16 +1,16 @@
 import json
 import unittest
 from unittest.mock import MagicMock
-from server.services.project_service import ProjectService, InvalidGeoJson, Project
+from server.services.project_admin_service import ProjectAdminService, InvalidGeoJson, Project
 
 
-class TestProject(unittest.TestCase):
+class TestProjectAdminService(unittest.TestCase):
 
     def test_cant_add_tasks_if_geojson_not_feature_collection(self):
         # Arrange
         invalid_feature = '{"coordinates": [[[[-4.0237, 56.0904], [-3.9111, 56.1715], [-3.8122, 56.098],' \
             '[-4.0237, 56.0904]]]], "type": "MultiPolygon"}'
-        test_project_service = ProjectService()
+        test_project_service = ProjectAdminService()
 
         # Act
         with self.assertRaises(InvalidGeoJson):
@@ -23,7 +23,7 @@ class TestProject(unittest.TestCase):
                                               '"MultiPolygon"}, "properties": {"x": 2402, "y": 1736, "zoom": 12}, "type":'
                                               '"Feature"}], "type": "FeatureCollection"}')
 
-        test_project_service = ProjectService()
+        test_project_service = ProjectAdminService()
         test_project = Project()
         test_project.create_draft_project('Test', MagicMock())
